parser_2 refactor, ft_strjoin bug fix

+++ libft/ft_strjoin.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strjoin.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: luinasci <luinasci@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jcologne <jcologne@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 16:00:43 by luinasci          #+#    #+#             */
-/*   Updated: 2025/05/05 17:09:41 by luinasci         ###   ########.fr       */
+/*   Updated: 2025/05/07 13:41:39 by jcologne         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,12 +25,12 @@ char	*ft_strjoin(char const *s1, char const *s2)
 	new_str = (char *)malloc(ft_strlen(s1) + ft_strlen(s2) + 1);
 	if (new_str == NULL)
 		return (NULL);
-	while (i < ft_strlen(s1))
+	while (i < (long unsigned int)ft_strlen(s1))
 	{
 		new_str[i] = s1[i];
 		i++;
 	}
-	while (j < ft_strlen(s2))
+	while (j < (long unsigned int)ft_strlen(s2))
 	{
 		new_str[i + j] = s2[j];
 		j++;
